fix: Address issues from code review
